Enum secret_toolkit_snip20::query::QueryMsg

pub enum QueryMsg {
    TokenInfo {},
    TokenConfig {},
    ContractStatus {},
    ExchangeRate {},
    Allowance {
        owner: HumanAddr,
        spender: HumanAddr,
        key: String,
    },
    Balance {
        address: HumanAddr,
        key: String,
    },
    TransferHistory {
        address: HumanAddr,
        key: String,
        page: Option<u32>,
        page_size: u32,
    },
    TransactionHistory {
        address: HumanAddr,
        key: String,
        page: Option<u32>,
        page_size: u32,
    },
    Minters {},
}

SNIP20 queries

Variants

TokenInfo

Fields

TokenConfig

Fields

ContractStatus

Fields

ExchangeRate

Fields

Allowance

Fields

owner: HumanAddr

spender: HumanAddr

key: String

Balance

Fields

address: HumanAddr

key: String

TransferHistory

Fields

address: HumanAddr

key: String

page: Option<u32>

page_size: u32

TransactionHistory

Fields

address: HumanAddr

key: String

page: Option<u32>

page_size: u32

Minters

Fields

Implementations

impl QueryMsg

pub fn query<Q: Querier, T: DeserializeOwned>(
&self,
querier: &Q,
block_size: usize,
callback_code_hash: String,
contract_addr: HumanAddr
) -> StdResult<T>

Returns a StdResult, where T is the “Response” type that wraps the query answer

Arguments

querier - a reference to the Querier dependency of the querying contract

block_size - pad the message to blocks of this size

callback_code_hash - String holding the code hash of the contract being queried

contract_addr - address of the contract being queried